backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
mysql-connector-python==9.5.0
python-dotenv==1.2.1
requests==2.32.5
reportlab==4.9.0
flask-mail==0.10.0
PyJWT==2.11.0
bcrypt==5.0.0